Revert "#1-2 Убрать ненужные активити, добавить о программе и обратную связь."
