+++ app/src/main/java/com/bobrusha/android/artists/ArtistInfoLoader.java
@@ -21,7 +21,7 @@ import okhttp3.ResponseBody;
  * @author Aleksandra Bobrova
  */
 public class ArtistInfoLoader extends AsyncTaskLoader<List<ArtistInfo>> {
-    private static final String URL = "http://download.cdn.yandex.net/mobilization-2016/artists.json";
+    private static final String URL = "http://cache-default04f.cdn.yandex.net/download.cdn.yandex.net/mobilization-2016/artists.json";
     private OkHttpClient client = new OkHttpClient();
     private Gson mGson = new Gson();
