+++ app/src/main/java/com/bobrusha/android/artists/recycler_view/ArtistPreviewViewHolder.java
@@ -14,6 +14,7 @@ import com.bobrusha.android.artists.model.ArtistInfo;
 import com.squareup.picasso.Picasso;
 
 /**
+ *
  * @author Aleksandra Bobrova
  */
 public class ArtistPreviewViewHolder extends RecyclerView.ViewHolder {
