--- app/src/main/res/layout/fragment_main.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v4.widget.SwipeRefreshLayout
-    android:id="@+id/refresh_layout"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/recycler_view_artists"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        />
-
-</android.support.v4.widget.SwipeRefreshLayout>
