--- app/src/main/res/layout/toolbar.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.Toolbar xmlns:android="http://schemas.android.com/apk/res/android"
-                                   xmlns:app="http://schemas.android.com/apk/res-auto"
-                                   android:id="@+id/toolbar"
-                                   android:layout_height="wrap_content"
-                                   android:layout_width="match_parent"
-                                   android:fitsSystemWindows="true"
-                                   android:minHeight="?attr/actionBarSize"
-                                   app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
-                                   android:background="?attr/colorPrimary"
-                                   app:popupTheme="@style/AppTheme.PopupOverlay">
-
-</android.support.v7.widget.Toolbar>
