--- app/src/main/res/menu/drawer_menu.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <group android:checkableBehavior="single">
-        <item
-            android:id="@+id/nav_home"
-            android:title="@string/nav_artist_option"/>
-        <item
-            android:id="@+id/nav_about_program"
-            android:title="@string/nav_about_program_option" />
-
-    </group>
-
-</menu>
